// src/app/interfaces/Supplier.ts
export interface Supplier
{
    id: string; 
    name: string;
    cnpj: string;
    openDate: string;
    email: string;
    phoneNumber: string;
    homeNumber: string;
    logradouro: string;
    complementResidential: string;
    neighborhood: string;
    city: string;
    state: string;
    active: boolean;
}